% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulepaths.R
\name{append.module.paths}
\alias{append.module.paths}
\title{Append/Insert a path into module.paths, similar to append()}
\usage{
append.module.paths(value, after = -1)
}
\arguments{
\item{value}{(relative) path to insert into \code{module.paths}}

\item{after}{location in \code{module.paths} to append, 0 for the beginning, defaults
to -1 which appends to the end}
}
\value{
Nothing is returned.
}
\description{
Beware, little error checking is done to see if the indexes are valid.
Note the item indexes are 1-based.
}
\examples{
# Inserts `../R/lib' as the second path to search for modules
append.module.paths('../R/lib', after = 1)
}

