% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findfirst.R
\name{find.first.R}
\alias{find.first.R}
\title{Returns the path of the first found instance of \code{file} in \code{module.path}.}
\usage{
find.first.R(file)
}
\arguments{
\item{file}{a string (or symbol) specifying the \code{file} to search for existance and readability in the current directory,
and if it cannot be found, searches for it in the list of directories specified by \code{module.paths}
and then through the set of paths with the file using a .R extension, if it was not originally
specified.}
}
\value{
A string consisting of the path the file was first found searching through module.paths.
}
\description{
A symbol maybe passed instead of a string for readability. If an expression is passed, it must return a string value.
}
\examples{
# Returns the path to the first found file according to module.paths
adder.path <- find.first.R('adder')
}

