% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcache.R
\name{get.module.cache}
\alias{get.module.cache}
\title{Returns the current file cache}
\usage{
get.module.cache()
}
\value{
environment containing the file cache.
}
\description{
Returns the current file cache
}
\examples{
cache <- get.module.cache()
}

