% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulepaths.R
\name{get.module.paths}
\alias{get.module.paths}
\title{Get existing collection of search paths of where to look for modules.}
\usage{
get.module.paths()
}
\value{
Vector of paths (strings) that specify folders where to search for module files, in order.
}
\description{
Get existing collection of search paths of where to look for modules.
}
\examples{
# Returns a copy of the current module.paths
paths <- get.module.paths()
}

