% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findfirst.R
\name{hide.not.found.warnings}
\alias{hide.not.found.warnings}
\title{Globally hide warnings when modules are not found}
\usage{
hide.not.found.warnings()
}
\value{
nothing is returned
}
\description{
This is only to be called when handling results manually. This will be overridden by a \code{warn.not.found}
parameter explicitly set by either \code{\link{lrequire}}, \code{\link{find.first.R}}.
}
\examples{
# Ensure warnings are not displayed when lrequire cannot find the module
hide.not.found.warnings()
}

