% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removefromcache.R
\name{remove.from.module.cache}
\alias{remove.from.module.cache}
\title{Removes module from cache, applying same logic as \code{\link{find.first.R}} to find and remove it}
\usage{
remove.from.module.cache(file)
}
\arguments{
\item{file}{name of a file, same as the one used in the \code{\link{lrequire}} method, that will be removed
from the cache, such that the next time the \code{file} is \code{\link{lrequire}}'d, it will be
read and executed.}
}
\value{
boolean value yielding success of removal from the cache
}
\description{
Removes module from cache, applying same logic as \code{\link{find.first.R}} to find and remove it
}
\examples{
remove.from.module.cache(variables)
}

