% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulepaths.R
\name{remove.module.paths}
\alias{remove.module.paths}
\title{Remove one or more paths from \code{module.paths}}
\usage{
remove.module.paths(index, ...)
}
\arguments{
\item{index}{index of item to be removed from path}

\item{...}{optional indexes of items to be removed from path}
}
\value{
Nothing is returned.
}
\description{
Beware, little error checking is done to see if the indexes are valid.
Note the item indexes are 1-based.
}
\examples{
# Removes the2nd and 4th items from module.paths
remove.module.paths(2, 4)
}

