% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetcache.R
\name{reset.module.cache}
\alias{reset.module.cache}
\title{Resets the module cache, ensuring files are loaded on next require}
\usage{
reset.module.cache()
}
\value{
Nothing is returned
}
\description{
Note the chace contains other hidden variables, kept in the cache (or environment). These are not removed,
and removing them will conflict with the ability of \code{\link{lrequire}} to perform properly.
}
\examples{
reset.module.cache()
}

