% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcache.R
\name{show.module.cache}
\alias{show.module.cache}
\title{Prints the current file cache}
\usage{
show.module.cache(all.names = FALSE)
}
\arguments{
\item{all.names}{a logical value. If TRUE, all object names are returned. If FALSE,
names which begin with a . are omitted.}
}
\value{
Nothing is returned, however, the contents of the module cache are printed to the standard output.
}
\description{
Prints the current file cache
}
\examples{
show.module.cache()
}

