% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findfirst.R
\name{show.not.found.warnings}
\alias{show.not.found.warnings}
\title{Globally show warnings when modules are not found}
\usage{
show.not.found.warnings()
}
\value{
nothing is returned
}
\description{
This is the default behaviour.
}
\details{
This is only to be called when handling results manually. This will be overridden by a \code{warn.not.found}
parameter explicitly set by either \code{\link{lrequire}} or \code{\link{find.first.R}}.
}
\examples{
# Ensure warnings are displayed when lrequire cannot find the module
show.not.found.warnings()
}

